+++ src/notebooks/deepnote/integrations/federatedAuth/federatedAuthTestHelpers.node.ts
@@ -1,0 +1,18 @@
+// Node-only test helpers; pulls in `googleOAuthProvider.node` for the loopback-flow + OAuth-provider tests.
+
+import {
+    buildBigQueryGoogleOAuthStrategy,
+    type BuildBigQueryGoogleOAuthStrategyParams,
+    createInMemoryPkceStore
+} from './googleOAuthProvider.node';
+
+export function buildTestStrategy(
+    overrides: Partial<BuildBigQueryGoogleOAuthStrategyParams> = {}
+): ReturnType<typeof buildBigQueryGoogleOAuthStrategy> {
+    return buildBigQueryGoogleOAuthStrategy({
+        clientId: 'cid',
+        clientSecret: 'cs',
+        store: createInMemoryPkceStore(),
+        ...overrides
+    });
+}
